requirements.txt
# Local package
-e .

# Side packages (general)
pandas==1.2.3
jupyter==1.0.0
scipy==1.6.2
numpy==1.19.5
setuptools==50.3.2
bootstrapped==0.0.2
matplotlib==3.4.3
seaborn==0.11.2

# Scraping
lxml==4.6.3
selenium==3.141.0
beautifulsoup4==4.9.3
tqdm==4.61.1

# NHL API stats extractor
requests==2.25.1

# xG model
scikit-learn==0.24.2
bs4==0.0.1
hockey_scraper==1.37.7
src==2.0
ThinkBayes2==2.0.0
tensorflow-gpu==2.5.0

